+++ kernel/ps2/ps2_keyboard.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <bolgenos-ng/ps2/device.hpp>
-#include <bolgenos-ng/keyboard.h>
+#include <bolgenos-ng/keyboard.hpp>
 
 #include "ps2_keyboard_sm.hpp"
